fix: change order_number from INTEGER to BIGINT for Postgres
The order_number format YYYYMMDD*1000+seq produces values like 20260223001
which exceeds Postgres INTEGER max (2,147,483,647). Change to BIGINT.

+++ apps/pos/server/db/pg-schema.sql
@@ -68,7 +68,7 @@ CREATE TABLE IF NOT EXISTS menu_items (
 CREATE TABLE IF NOT EXISTS orders (
   id SERIAL PRIMARY KEY,
   tenant_id TEXT NOT NULL DEFAULT current_setting('app.tenant_id', true),
-  order_number INTEGER NOT NULL,
+  order_number BIGINT NOT NULL,
   employee_id INTEGER NOT NULL REFERENCES employees(id),
   status TEXT DEFAULT 'pending',
   subtotal NUMERIC(10,2),
